% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{conditional_intensity}
\alias{conditional_intensity}
\title{Compute Conditional Intensity at Each Event}
\usage{
conditional_intensity(hawkes, parameters)
}
\arguments{
\item{hawkes}{A \code{hawkes} object.}

\item{parameters}{Named list containing background, triggering, spatial, and temporal
parameters. Values usually represent current estimates.}
}
\value{
A numeric vector.
}
\description{
Compute Conditional Intensity at Each Event
}
\examples{
spatial_region <- create_rectangular_sf(0, 10, 0, 10)

params <- list(
  background_rate = list(intercept = -4),
  triggering_rate = 0.75,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2)
)
hawkes <- rHawkes(
  params = params,
  time_window = c(0, 50),
  spatial_region = spatial_region,
  background_process = ~ 1
)
conditional_intensity(hawkes, params)

spatial_region <- create_rectangular_sf(0, 10, 0, 10)

params <- list(
  background_rate = list(intercept = -4,
                         event_type = c(a = 1, b = .25, c = .5)),
  triggering_rate = matrix(c(.4, .15, .05,
                             .2, .05, .02,
                             .2, .05, .20),
                           nrow = 3,
                           dimnames = list(c("a", "b", "c"), c("a", "b", "c"))),
  spatial = list(mean = 0, sd = 0.1),
  temporal = list(rate = 2)
)
(hawkes <- rHawkes(
  params = params,
  time_window = c(0, 50),
  spatial_region = spatial_region,
  background_process = ~ 1 + mark(event_type),
  spatial_burnin = 1
))

conditional_intensity(hawkes, params)
}
