% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rHawkes.R
\name{create_rectangular_sf}
\alias{create_rectangular_sf}
\title{Create Rectangular SF Object}
\usage{
create_rectangular_sf(xmin, xmax, ymin, ymax, crs = NA)
}
\arguments{
\item{xmin}{horizontal minimum of spatial region}

\item{xmax}{horizontal maximum of spatial region}

\item{ymin}{vertical minimum of spatial region}

\item{ymax}{vertical maximum of spatial region}

\item{crs}{coordinate reference system. Defaults to NA if unused.}
}
\value{
an sf object with a rectangular region
}
\description{
Create Rectangular SF Object
}
\examples{
create_rectangular_sf(0,10,0,10)
}
