% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{dexp_spatial}
\alias{dexp_spatial}
\title{Density of events with exponential(rate) distance in random direction from the origin}
\usage{
dexp_spatial(x, rate)
}
\arguments{
\item{x}{numeric vector of x coordinates}

\item{rate}{rate parameter of the exponential distribution}

\item{y}{numeric vector of y coordinates}
}
\value{
numeric vector of densities at each (x, y) point
}
\description{
Density of events with exponential(rate) distance in random direction from the origin
}
\examples{
dexp_spatial(c(1, 0), c(0, 1), rate = 2)
}
