% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rHawkes.R
\name{.blank_hawkes}
\alias{.blank_hawkes}
\title{Create an empty hawkes object}
\usage{
.blank_hawkes(params = NULL, region = NULL)
}
\arguments{
\item{params}{A named list of lists containing the values for the background rate, triggering ratio, spatial parameters in a named list, and temporal parameters in a named list.}

\item{region}{A list containing the spatial and temporal windows of the form list(x = c(xmin, xmax), y = c(ymin, ymax), t = c(tmin, tmax)).}
}
\value{
An empty hawkes object.
}
\description{
Create an empty hawkes object
}
