% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.sanity_check}
\alias{.sanity_check}
\title{(Internal) Check to make sure necessary attributes are in the Hawkes object}
\usage{
.sanity_check(hawkes)
}
\arguments{
\item{hawkes}{A hawkes object}
}
\value{
TRUE
}
\description{
(Internal) Check to make sure necessary attributes are in the Hawkes object
}
