% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximization_functions.R
\name{.spatial_parameter_likelihood}
\alias{.spatial_parameter_likelihood}
\title{Spatial triggering kernel parameter likelihood}
\usage{
.spatial_parameter_likelihood(p, hawkes, parent_est_mat, x_diff, y_diff)
}
\arguments{
\item{p}{A named list of parameters for the spatial kernel.}

\item{hawkes}{A \code{hawkes} object.}

\item{parent_est_mat}{Matrix of estimated parent probabilities.}

\item{x_diff}{A matrix of pairwise x-coordinate differences.}

\item{y_diff}{A matrix of pairwise y-coordinate differences.}
}
\value{
The negative log-likelihood contribution of the spatial triggering parameters.
}
\description{
Computes the (negative) expected complete-data log-likelihood for spatial triggering kernel parameters, assuming separable spatial kernels.
}
\keyword{internal}
