% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_update_functions.R
\name{.spatial_update_gaussian}
\alias{.spatial_update_gaussian}
\title{Gaussian spatial parameter update}
\usage{
.spatial_update_gaussian()
}
\value{
A list of updated sd estimate
}
\description{
This is an internal helper and not
meant to be called directly by users.
}
\keyword{internal}
