% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximization_functions.R
\name{.temporal_parameter_likelihood}
\alias{.temporal_parameter_likelihood}
\title{Temporal kernel likelihood contribution}
\usage{
.temporal_parameter_likelihood(
  p,
  hawkes,
  parent_est_mat,
  time_diff,
  triggering_rate
)
}
\arguments{
\item{p}{A named list of parameters for the temporal kernel.}

\item{hawkes}{A \code{hawkes} object.}

\item{parent_est_mat}{Matrix of estimated parent probabilities.}

\item{time_diff}{A matrix of pairwise time differences.}
}
\value{
The negative log-likelihood contribution of the temporal parameters.
}
\description{
Evaluates the negative expected complete-data log-likelihood for temporal triggering
kernels.
}
\keyword{internal}
