% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_update_functions.R
\name{.temporal_update_exponential}
\alias{.temporal_update_exponential}
\title{Exponential temporal parameter update}
\usage{
.temporal_update_exponential(hawkes)
}
\value{
A list of the updated rate estimate
}
\description{
This is an internal helper and not
meant to be called directly by users.
}
\keyword{internal}
