% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.unpack_hawkes}
\alias{.unpack_hawkes}
\title{(Internal) Unpack all Attributes in Hawkes Object}
\usage{
.unpack_hawkes(hawkes)
}
\arguments{
\item{hawkes}{A hawkes object}
}
\value{
Loads attributes to function environment
}
\description{
(Internal) Unpack all Attributes in Hawkes Object
}
