% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{dpower_law}
\alias{dpower_law}
\title{Density function for power law (Lomax)}
\usage{
dpower_law(x, shape = 2, scale = 1)
}
\arguments{
\item{x}{numeric vector}

\item{shape}{shape parameter for the Lomax distribution}

\item{scale}{scale parameter for the Lomax distribution}
}
\description{
Density function for power law (Lomax)
}
\examples{
dpower_law(1, shape = 2, scale = 1)
dpower_law(matrix(c(1:48, NA), nrow = 7), shape = 2, scale = 1)
}
