% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_background_covariates.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example Simulated Hawkes Data}
\format{
A \code{hawkes} object with a covariate named \code{z}.
}
\usage{
example_data
}
\description{
A simple \code{hawkes} object simulated from the \code{example_background_covariates}
spatial region.
}
\examples{

data(example_data)
plot_hawkes(example_data)

inits <- list(
  background_rate = list(intercept = -5, z = .5),
  triggering_rate = 0.5,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2),
  fixed = list(spatial = "mean")
)
hawkes_mle(example_data, inits = inits, boundary = .5)
}
\keyword{datasets}
