% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{full_log_likelihood}
\alias{full_log_likelihood}
\title{Compute Full Log-Likelihood}
\usage{
full_log_likelihood(hawkes, parameters)
}
\arguments{
\item{hawkes}{A hawkes object}

\item{parameters}{A named list of lists containing the values for the background rate, triggering ratio, spatial parameters in a named list, and temporal parameters in a named list. Note that these values are of the same form as the true values in the Hawkes object but are often estimates passed to the function.}
}
\value{
A numeric vector.
}
\description{
Compute Full Log-Likelihood
}
\examples{
region <- list(x = c(0,10), y = c(0,10), t = c(0,50))

params <- list(background_rate = -4,triggering_rate = 0.5,spatial = list(mean = 0, sd = 0.5),temporal = list(rate = 2))
hawkes <- rHawkes(params, region)
full_log_likelihood(hawkes, params)

params <- list(background_rate = -4,triggering_rate = 0.5,spatial = list(min = -1, max = 1),temporal = list(shape = 3, rate = 1))
hawkes <- rHawkes(params, region, spatial_family = "Uniform", temporal_family = "Gamma")
full_log_likelihood(hawkes, params)

params <- list(background_rate = -4,triggering_rate = 0.5,spatial = list(rate = 5),temporal = list(rate = 2))
hawkes <- rHawkes(params, region, spatial_family = "Exponential", temporal_family = "Exponential")
full_log_likelihood(hawkes, params)
}
