% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_object.R
\name{new_hawkes}
\alias{new_hawkes}
\title{Create hawkes object}
\usage{
new_hawkes(
  data = NULL,
  params = NULL,
  region = NULL,
  spatial_kernel = NULL,
  temporal_kernel = NULL,
  cov_map = NULL
)
}
\arguments{
\item{data}{A dataframe containing the event locations in the columns x, y, and t.}

\item{params}{A named list of lists containing the values for the background rate, triggering ratio, spatial parameters in a named list, and temporal parameters in a named list. Defaults to NULL if not used.}

\item{region}{A list containing the spatial and temporal windows of the form list(x = c(xmin, xmax), y = c(ymin, ymax), t = c(tmin, tmax)).}

\item{spatial_kernel}{A spatial triggering kernel function to generate data from. Defaults to NULL if not used.}

\item{temporal_kernel}{A spatial triggering kernel function to generate data from. Defaults to NULL if not used.}

\item{cov_map}{An sf object containing spatial polygons with associated covariate values named X1, X2, .... Defaults to NULL if not used.}
}
\value{
A hawkes object containing a tibble with the events.
}
\description{
Create hawkes object
}
\examples{
new_hawkes()
}
