% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scratch.R
\docType{data}
\name{params}
\alias{params}
\title{#' Title
#'
#' @param i
#' @param parent_mat
#'
#' @returns
#' @export
#'
#' @examples
sample_descendants <- function(i, parent_mat, x_diff, y_diff, time_diff) {
n <- nrow(parent_mat)}
\format{
An object of class \code{list} of length 5.
}
\usage{
params
}
\description{
exp_desc <- sum(parent_mat) / n
(n_desc <- rpois(1, exp_desc))
}
\details{
if (n_desc == 0) {
return(data.frame())
}

mat <- matrix(1:n^2, nrow = n)
}
\section{Sample indices of descendant differences}{
idx <- sample(seq_along(mat), size = n_desc, prob = as.vector(parent_mat), replace = TRUE)
}

\section{Convert back to row/col indices}{

}

\section{Create matrices with the differences in location and times between each point}{
x_diff <- outer(hawkes$x, hawkes$x, \code{-})
x_diff\link{upper.tri(x_diff)} <- 0
y_diff <- outer(hawkes$y, hawkes$y, \code{-})
y_diff\link{upper.tri(y_diff)} <- 0

time_diff <- outer(hawkes$t, hawkes$t, \code{-})
time_diff\link{upper.tri(time_diff)} <- 0

n <- nrow(hawkes)
}

\section{Sample the number of background events and events themselves as a Poisson random variable}{

}

\section{with mean equal to the sum of the diagonal of the parent matrix}{

}

\keyword{datasets}
