% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{pexp_spatial}
\alias{pexp_spatial}
\title{CDF of events with exponential(rate) distance in random direction from the origin}
\usage{
pexp_spatial(q, rate)
}
\arguments{
\item{q}{a nx2 matrix of coordinates}

\item{rate}{rate parameter of the exponential distribution}
}
\value{
numeric vector of cumulative probabilities at each (x, y) point
}
\description{
CDF of events with exponential(rate) distance in random direction from the origin
}
\examples{
pexp_spatial(matrix(c(1, 0, 0, 1), ncol = 2), rate = 2)
}
