% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{ppower_law}
\alias{ppower_law}
\title{CDF for power law (Lomax)}
\usage{
ppower_law(q, shape = 2, scale = 1, lower.tail = TRUE)
}
\arguments{
\item{q}{numeric vector}

\item{shape}{shape parameter for the Lomax distribution}

\item{scale}{scale parameter for the Lomax distribution}

\item{lower.tail}{logical; if TRUE (default), probabilities are $P\link{X< x}$ otherwise, $P\link{X>x}$.}
}
\value{
a numeric vector of cumulative probabilities at each value of q
}
\description{
CDF for power law (Lomax)
}
\examples{
ppower_law(1, shape = 2, scale = 1)
}
