% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_object.R
\name{print.hawkes}
\alias{print.hawkes}
\title{Title}
\usage{
\method{print}{hawkes}(x, n = 10, ...)
}
\arguments{
\item{x}{a hawkes object to be printed}

\item{n}{the number of events to print}

\item{...}{arguments passed to or from other methods}
}
\value{
The input hawkes object, invisibly.
}
\description{
Title
}
