% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_fit.R
\name{print.hawkes_fit}
\alias{print.hawkes_fit}
\title{Print hawkes fit object}
\usage{
\method{print}{hawkes_fit}(x)
}
\arguments{
\item{x}{a hawkes fit object to be printed}
}
\value{
The input hawkes fit object, invisibly.
}
\description{
Print hawkes fit object
}
