% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scratch.R
\name{recursive_sample_descendants}
\alias{recursive_sample_descendants}
\title{Title}
\usage{
recursive_sample_descendants(
  i,
  hawkes,
  parent_mat,
  boundary,
  x_diff,
  y_diff,
  time_diff
)
}
\arguments{
\item{time_diff}{}
}
\description{
Title
}
