% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{rexp_spatial}
\alias{rexp_spatial}
\title{Sample events with exponential(rate) distance in random direction from the origin}
\usage{
rexp_spatial(n, rate)
}
\arguments{
\item{n}{number of observations}

\item{rate}{vector of rates}
}
\value{
a dataframe of points located exp(rate) away from the origin
}
\description{
Sample events with exponential(rate) distance in random direction from the origin
}
\examples{
rexp_spatial(5, 2)
}
