% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{rpower_law}
\alias{rpower_law}
\title{Sample events with power law distance in time}
\usage{
rpower_law(n, shape = 2, scale = 1)
}
\arguments{
\item{n}{number of observations}

\item{shape}{shape parameter for the Lomax distribution}

\item{scale}{scale parameter for the Lomax distribution}
}
\value{
a numeric vector of values of the Lomax distribution
}
\description{
Sample events with power law distance in time
}
\examples{
rpower_law(5, shape = 2, scale = 1)
}
