% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_bootstrap.R
\name{sample_clusters}
\alias{sample_clusters}
\title{Resample Data by EM Algorithm Clusters}
\usage{
sample_clusters(hawkes, parent_mat, boundary = NULL)
}
\arguments{
\item{hawkes}{A \code{hawkes} object to resample.}

\item{parent_mat}{Parent probabilities estimated with \code{parent_est()}.}

\item{boundary}{Optional boundary width used for edge correction.}
}
\value{
A resampled \code{hawkes} object.
}
\description{
Resample Data by EM Algorithm Clusters
}
\examples{
params <- list(
  background_rate = list(intercept = -4.5, z = 1),
  triggering_rate = 0.5,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2),
  fixed = list(spatial = "mean")
)
data("example_background_covariates")
hawkes <- rHawkes(
  params,
  c(0, 50),
  example_background_covariates,
  covariate_columns = "z",
  spatial_burnin = 1
)
est <- hawkes_mle(hawkes, inits = params, boundary = 1)
parent_mat <- parent_est(hawkes, est)
sample_clusters(hawkes, parent_mat, boundary = c(.5,3))

}
