% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stHawkesTools-package.R
\docType{package}
\name{stHawkesTools-package}
\alias{stHawkesTools}
\alias{stHawkesTools-package}
\title{stHawkesTools: Provide Tools for Bootstrapping General Spatio-Temporal Hawkes Processes}
\description{
stHawkesTools is an R package for simulating, estimating, and analyzing spatio-temporal Hawkes processes. It supports flexible triggering kernels beyond those implemented in existing packages like ETAS, and provides simulation tools tailored for spatio-temporal settings. Parameter estimation is performed using the EM algorithm for maximum likelihood inference. The package also offers confidence intervals based on both asymptotic theory and a range of bootstrap methods. Additional utilities are included to support diagnostics and model-based analysis of spatio-temporal point process data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cfox20/stHawkesTools}
  \item \url{https://cfox20.github.io/stHawkesTools/}
  \item Report bugs at \url{https://github.com/cfox20/stHawkesTools/issues}
}

}
\author{
\strong{Maintainer}: Caleb Fox \email{caleb_fox2@ebaylor.edu} (\href{https://orcid.org/0000-0002-3992-6746}{ORCID})

}
\keyword{internal}
