% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_fit.R
\name{summary.hawkes_fit}
\alias{summary.hawkes_fit}
\title{Print hawkes fit object}
\usage{
\method{summary}{hawkes_fit}(object, ...)
}
\arguments{
\item{object}{a hawkes fit object to be printed}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print hawkes fit object
}
