% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{temporal_conditional_intensity}
\alias{temporal_conditional_intensity}
\title{Compute the conditional intensity at a location}
\usage{
temporal_conditional_intensity(
  hawkes,
  parameters,
  coordinates,
  step = 0.1,
  time_window = NULL
)
}
\arguments{
\item{hawkes}{A \code{hawkes} object.}

\item{parameters}{Named list containing background, triggering, spatial, and temporal
parameters.}

\item{coordinates}{Numeric vector of length two giving the evaluation location.}

\item{step}{Grid step size used to build the temporal evaluation grid.}

\item{time_window}{Optional numeric vector of length two giving the time interval over which to
evaluate the intensity.}
}
\value{
A numeric vector.
}
\description{
Compute the conditional intensity at a location
}
\examples{
spatial_region <- create_rectangular_sf(0, 10, 0, 10)

params <- list(
  background_rate = list(intercept = -4),
  triggering_rate = 0.75,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2)
)
hawkes <- rHawkes(
  params = params,
  time_window = c(0, 50),
  spatial_region = spatial_region,
  background_process = ~ 1
)
temporal_conditional_intensity(hawkes, params, c(5, 5))
}
