% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_object.R
\name{as_hawkes}
\alias{as_hawkes}
\title{Convert an Object to Type Hawkes}
\usage{
as_hawkes(
  data,
  region,
  params = NULL,
  spatial_kernel = NULL,
  temporal_kernel = NULL,
  cov_map = NULL
)
}
\arguments{
\item{data}{A dataframe containing the event locations in the columns x, y, and t.}

\item{region}{A list containing the spatial and temporal windows of the form list(x = c(xmin, xmax), y = c(ymin, ymax), t = c(tmin, tmax)).}

\item{params}{A named list of lists containing the values for the background rate, triggering ratio, spatial parameters in a named list, and temporal parameters in a named list. Defaults to NULL if not used.}

\item{spatial_kernel}{A spatial triggering kernel function to generate data from. Defaults to NULL if not used.}

\item{temporal_kernel}{A spatial triggering kernel function to generate data from. Defaults to NULL if not used.}

\item{cov_map}{An sf object containing spatial polygons with associated covariate values named X1, X2, .... Defaults to NULL if not used.}
}
\value{
A hawkes object.
}
\description{
Convert an Object to Type Hawkes
}
\examples{
set.seed(123)

df <- data.frame(
  x = runif(100, 0, 10),
  y = runif(100, 0, 10),
  t = runif(100, 0, 50)
)

# Convert to hawkes object
region <- list(x = c(0, 10), y = c(0, 10), t = c(0, 50))
hawkes_df <- as_hawkes(df, region)
print(hawkes_df)
}
