% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_object.R
\name{as_hawkes}
\alias{as_hawkes}
\title{Convert an Object to Type Hawkes}
\usage{
as_hawkes(
  data,
  time_window,
  spatial_region,
  spatial_family,
  temporal_family,
  covariate_columns = NULL
)
}
\arguments{
\item{data}{A dataframe containing the event locations in the columns x, y, and t or an sf object containing the event locations in geometry and the event times in a column named t.}

\item{time_window}{A numeric vector of length 2 specifying the time window.}

\item{spatial_region}{An sf object defining the spatial region and covariate regions.}

\item{spatial_family}{A spatial triggering kernel function to generate data from. Defaults to NULL if not used.}

\item{temporal_family}{A spatial triggering kernel function to generate data from. Defaults to NULL if not used.}

\item{covariate_columns}{A character vector containing the names of covariates columns to be used for the model. Defaults to NULL if not used}
}
\value{
A hawkes object.
}
\description{
Convert an Object to Type Hawkes
}
\examples{
df <- data.frame(
  x = runif(100, 0, 10),
  y = runif(100, 0, 10),
  t = runif(100, 0, 50)
)

# Convert to hawkes object
spatial_region <- create_rectangular_sf(0,10,0,10)
hawkes_df <- as_hawkes(df, c(0,50), spatial_region, spatial_family = "Gaussian", temporal_family = "Exponential")
print(hawkes_df)

}
