% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_object.R
\name{as_hawkes}
\alias{as_hawkes}
\title{Convert an Object to Type Hawkes}
\usage{
as_hawkes(
  data,
  time_window,
  spatial_region,
  spatial_family,
  temporal_family,
  params = NULL,
  cov_map = NULL,
  X = NULL
)
}
\arguments{
\item{data}{A dataframe containing the event locations in the columns x, y, and t. Defaults to NULL if not used.}

\item{spatial_family}{A spatial triggering kernel function to generate data from. Alternatively, a list can be provided to designate a custom kernel. The list must contain the objects named spatial_pdf, spatial_cdf, and spatial_sampler. They should follow the format of the dnorm, pnorm, and rnorm functions and the parameters must match the names. Defaults to NULL if not used.}

\item{temporal_family}{A spatial triggering kernel function to generate data from. Alternatively, a list can be provided to designate a custom kernel. The list must contain the objects named temporal_pdf, temporal_cdf, and temporal_sampler. They should follow the format of the dnorm, pnorm, and rnorm functions and the parameters must match the names. Defaults to NULL if not used.}

\item{params}{A named list of lists containing the values for the background rate, triggering ratio, spatial parameters in a named list, and temporal parameters in a named list. Defaults to NULL if not used.}

\item{cov_map}{An sf object containing spatial polygons with associated covariate values named X1, X2, .... Defaults to NULL if not used.}

\item{X}{A dataframe with nrow(data) rows containing the covariates observed at each event. Defaults to NULL if not used.}

\item{region}{A list containing the spatial and temporal windows of the form list(x = c(xmin, xmax), y = c(ymin, ymax), t = c(tmin, tmax)). Defaults to NULL if not used.}
}
\value{
A hawkes object.
}
\description{
Convert an Object to Type Hawkes
}
\examples{
set.seed(123)

df <- data.frame(
  x = runif(100, 0, 10),
  y = runif(100, 0, 10),
  t = runif(100, 0, 50)
)

# Convert to hawkes object
spatial_region <- create_rectangular_sf(0,10,0,10)
hawkes_df <- as_hawkes(df, c(0,50), spatial_region, spatial_family = "Gaussian", temporal_family = "Exponential")
print(hawkes_df)

}
