% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_bootstrap.R
\name{block_bootstrap}
\alias{block_bootstrap}
\title{Block bootstrap for Hawkes MLE confidence intervals}
\usage{
block_bootstrap(
  hawkes,
  est,
  B,
  num_blocks,
  alpha = 0.05,
  parallel = FALSE,
  max_iters = 500,
  boundary = NULL
)
}
\arguments{
\item{hawkes}{A \code{hawkes} object.}

\item{est}{A \code{hawkes_est} object containing parameter estimates for \code{hawkes}.}

\item{B}{Number of bootstrap iterations.}

\item{num_blocks}{Number of blocks sampled during each bootstrap iteration.}

\item{alpha}{Type I error rate for the confidence intervals. Defaults to 0.05.}

\item{parallel}{Logical flag for \code{furrr} based parallel computation.}

\item{max_iters}{Maximum EM iterations per fit. Defaults to 500.}

\item{boundary}{Optional boundary width used for edge correction.}
}
\value{
A tibble of bootstrap estimates.
}
\description{
Block bootstrap for Hawkes MLE confidence intervals
}
\examples{
params <- list(
  background_rate = list(intercept = -4.5, z = 1),
  triggering_rate = 0.5,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2),
  fixed = list(spatial = "mean")
)
data("example_background_covariates")
hawkes <- rHawkes(
  params,
  c(0, 50),
  example_background_covariates,
  covariate_columns = "z",
  spatial_burnin = 1
)
est <- hawkes_mle(hawkes, inits = params, boundary = 1)

block_bootstrap(
  hawkes,
  est,
  B = 2,
  #B = 1000,
  num_blocks = 25,
  alpha = 0.05,
  parallel = FALSE,
  boundary = 1
)

}
