% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_bootstrap.R
\name{cluster_bootstrap}
\alias{cluster_bootstrap}
\title{Block Bootstrap for Confidence Intervals of Hawkes MLEs}
\usage{
cluster_bootstrap(
  hawkes,
  est,
  B,
  alpha = 0.05,
  parallel = FALSE,
  max_iters = 500,
  boundary = NULL
)
}
\arguments{
\item{hawkes}{A \code{hawkes} object}

\item{est}{A \code{hawkes_est} object containing the parameter estimates of \code{hawkes}}

\item{B}{number of bootstrap iterations}

\item{alpha}{type-1 error rate for constructing confidence intervals. Defaults to .05 if unused}

\item{parallel}{a logical specifying if parallel computation should be used. Parallel computation is implemented with the \code{furrr} package.}

\item{max_iters}{maximum number of iterations to use in maximum likelihood estimation. Defaults to 500 if unused.}

\item{boundary}{size of boundary to use for border correction. Defaults to NULL if unused}
}
\value{
A \code{hawkes} object.
}
\description{
Block Bootstrap for Confidence Intervals of Hawkes MLEs
}
\examples{
params <- list(background_rate = list(intercept = -4.5, X1 = 1, X2 = 1),triggering_rate = 0.5,spatial = list(mean = 0, sd = .25),temporal = list(rate = 2), fixed = list(spatial = "mean"))
data("example_background_covariates")
hawkes <- rHawkes(params, c(0,50), example_background_covariates, covariate_columns = c("X1", "X2"), spatial_burnin = 1)
est <- hawkes_mle(hawkes, inits = params, boundary = 1)

future::plan(future::multisession, workers = future::availableCores())

cluster_bootstrap(hawkes, est, B = 2, alpha = .05, parallel = TRUE, boundary = c(.5,3))

future::plan(future::sequential)
}
