% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_bootstrap.R
\name{cluster_bootstrap}
\alias{cluster_bootstrap}
\title{Block Bootstrap for Confidence Intervals of Hawkes MLEs}
\usage{
cluster_bootstrap(
  hawkes,
  est,
  B,
  alpha,
  parallel = FALSE,
  seed = NULL,
  max_iters = 500,
  boundary = NULL,
  t_burnin = 10,
  s_burnin = 0
)
}
\arguments{
\item{hawkes}{A \code{hawkes} object}

\item{block_length_t}{A numeric value to set the length of blocks to wrap the process.}
}
\value{
A \code{hawkes} object.
}
\description{
Block Bootstrap for Confidence Intervals of Hawkes MLEs
}
\examples{
region <- list(x = c(0,10), y = c(0,10), t = c(0,150))

params <- list(background_rate = list(intercept = -4),triggering_rate = 0.5,spatial = list(mean = 0, sd = 0.1),temporal = list(rate = 2), fixed = list(spatial = "mean", temporal = NULL))
hawkes <- rHawkes(params, region)
est <- hawkes_mle(hawkes, inits = params, boundary = c(.5, 3))
cluster_bootstrap(hawkes, est, 50, alpha = 0.05)

data("example_background_covariates")
params <- list(background_rate = list(intercept = -4.5, X1 = 1, X2 = 1, X3 = 1),triggering_rate = 0.5,spatial = list(mean = 0, sd = .1),temporal = list(rate = 2), fixed = list(spatial = "mean"))
hawkes <- rHawkes(params, region, spatial_family = "Gaussian", temporal_family = "Exponential", cov_map = example_background_covariates)
est <- hawkes_mle(hawkes, inits = params, boundary = c(.5, 3))

future::plan(future::multisession, workers = future::availableCores())

cluster_bootstrap(hawkes, est, B = 50, alpha = .05, parallel = TRUE, seed = 123, boundary = c(.5,3))

future::plan(future::sequential)
}
