% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rHawkes.R
\name{create_rectangular_sf}
\alias{create_rectangular_sf}
\title{Create Rectangular SF Object}
\usage{
create_rectangular_sf(
  xmin,
  xmax,
  ymin,
  ymax,
  covariates = NULL,
  n_grid = c(1, 1),
  crs = NA
)
}
\arguments{
\item{xmin}{horizontal minimum of spatial region}

\item{xmax}{horizontal maximum of spatial region}

\item{ymin}{vertical minimum of spatial region}

\item{ymax}{vertical maximum of spatial region}

\item{covariates}{a matrix, data.frame, or tibble containing columns specifying covariate values for the grid.
The number of rows must equal \code{n_grid[1]*n_grid[2]}. The values fill in the region
left to right, then bottom to top.}

\item{n_grid}{numeric vector of length 2 specifying dimensions for the grid
of covariate regions. e.g. \code{n_grid = c(10,20)}}

\item{crs}{coordinate reference system. Defaults to NA if unused.}
}
\value{
an sf object with a rectangular region
}
\description{
Create Rectangular SF Object
}
\examples{
create_rectangular_sf(0,10,0,10)
}
