% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{dexp_spatial}
\alias{dexp_spatial}
\title{Density of events with exponential(rate) distance in random direction from the origin}
\usage{
dexp_spatial(x, rate)
}
\arguments{
\item{x}{a nx2 matrix of coordinates}

\item{rate}{rate parameter of the exponential distribution}
}
\value{
numeric vector of densities at each (x, y) point
}
\description{
Density of events with exponential(rate) distance in random direction from the origin
}
\examples{
dexp_spatial(matrix(c(1, 0, 0, 1), ncol = 2), rate = 2)
}
