% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximization_functions.R
\name{.safe_log}
\alias{.safe_log}
\title{Safe logarithm to avoid log(0)}
\usage{
.safe_log(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector equal to \code{log(pmax(x, .Machine$double.eps))}.
}
\description{
Safe logarithm with lower bound
}
\keyword{internal}
