% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximization_functions.R
\name{.temporal_parameter_likelihood}
\alias{.temporal_parameter_likelihood}
\title{Temporal triggering kernel parameter likelihood}
\usage{
.temporal_parameter_likelihood(p, hawkes, parent_est_mat, time_diff)
}
\arguments{
\item{p}{A named list of parameters for the temporal kernel.}

\item{hawkes}{A \code{hawkes} object.}

\item{parent_est_mat}{Matrix of estimated parent probabilities.}

\item{time_diff}{A matrix of pairwise time differences.}
}
\value{
The negative log-likelihood contribution of the temporal triggering parameters.
}
\description{
Computes the (negative) expected complete-data log-likelihood for temporal triggering kernel parameters.
}
\keyword{internal}
