% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.vector_input_log_likelihood}
\alias{.vector_input_log_likelihood}
\title{Wrapper for likelihood for numerical optimization}
\usage{
.vector_input_log_likelihood(par_vec, hawkes, param_template)
}
\arguments{
\item{par_vec}{a numeric vector of the parameter initial values}

\item{hawkes}{a hawkes object}

\item{param_template}{a template to reconstruct the parameter object}
}
\description{
Wrapper for likelihood for numerical optimization
}
\keyword{internal}
