% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_likelihood_estimation.R
\name{est_params}
\alias{est_params}
\title{Compute Parameter Estimate Updates}
\usage{
est_params(
  hawkes,
  parameters,
  parent_est_mat,
  boundary = NULL,
  fixed_spatial = NULL,
  fixed_temporal = NULL
)
}
\arguments{
\item{hawkes}{A \code{hawkes} object}

\item{parameters}{A named list of lists containing the values for the background rate, triggering ratio, spatial parameters in a named list, and temporal parameters in a named list. Note that these values are of the same form as the true values in the Hawkes object but are often estimates passed to the function. If some parameters in the kernel function should be fixed in estimation, pass their names as another named list within params as fixed$spatial = c("mean").}

\item{parent_est_mat}{A matrix produced from the parent_est_mat() function.}

\item{boundary}{A boundary region to correct for the bpundary bias. Defaults to NULL if not used.}
}
\value{
A named list of the form of parameters with udpated parameter estimates.
}
\description{
Compute Parameter Estimate Updates
}
\examples{
region <- list(x = c(0,10), y = c(0,10), t = c(0,50))

params <- list(background_rate = list(intercept = -4),triggering_rate = 0.5,spatial = list(mean = 0, sd = 0.5),temporal = list(rate = 2), fixed = list(spatial = "mean", temporal = NULL))
hawkes <- rHawkes(params, region)
parent_est_mat <- parent_est(hawkes, params)
est_params(hawkes, params, parent_est_mat)

}
