% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_background_covariates.R
\docType{data}
\name{example_background_covariates}
\alias{example_background_covariates}
\title{Example Background Covariate sf Object}
\format{
An \code{sf} object with geometries and covariates named X1, X2, and X3
}
\source{
Bounding box and covariates derived from U.S. Census Bureau data accessed via \code{tidycensus}
}
\usage{
example_background_covariates
}
\description{
A simple \code{sf} object with a rectangular spatial region and one covariate column,
useful for simulating background intensity in Hawkes processes.
}
\examples{
library("sf")

data(example_background_covariates)
plot(example_background_covariates)
}
\keyword{datasets}
