% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_bootstrap.R
\name{extend_data_t_only}
\alias{extend_data_t_only}
\title{Wrap data for block sampling}
\usage{
extend_data_t_only(hawkes, block_length_t)
}
\arguments{
\item{hawkes}{A \code{hawkes} object}

\item{block_length_t}{A numeric value to set the length of blocks to wrap the process.}
}
\value{
A \code{hawkes} object.
}
\description{
Wrap data for block sampling
}
\examples{
params <- list(background_rate = list(intercept = -4.5, X1 = 1, X2 = 1),triggering_rate = 0.5,spatial = list(mean = 0, sd = .25),temporal = list(rate = 2), fixed = list(spatial = "mean"))
data("example_background_covariates")
hawkes <- rHawkes(params, c(0,50), example_background_covariates, covariate_columns = c("X1", "X2"), spatial_burnin = 1)
extend_data_t_only(hawkes, 5)

}
