% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_bootstrap.R
\name{extend_data_t_only}
\alias{extend_data_t_only}
\title{Wrap data for block sampling}
\usage{
extend_data_t_only(hawkes, block_length_t)
}
\arguments{
\item{hawkes}{A \code{hawkes} object}

\item{block_length_t}{A numeric value to set the length of blocks to wrap the process.}
}
\value{
A \code{hawkes} object.
}
\description{
Wrap data for block sampling
}
\examples{
region <- list(x = c(0,10), y = c(0,10), t = c(0,50))

params <- list(background_rate = list(intercept = -4),triggering_rate = 0.5,spatial = list(mean = 0, sd = 0.1),temporal = list(rate = 2), fixed = list(spatial = "mean", temporal = NULL))
hawkes <- rHawkes(params, region)
extend_data_t_only(hawkes, 5)

}
