% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_object.R
\name{hawkes}
\alias{hawkes}
\title{Constructor for a hawkes object}
\usage{
hawkes(
  data = NULL,
  params = NULL,
  time_window = NULL,
  spatial_region = NULL,
  spatial_family = NULL,
  temporal_family = NULL,
  covariate_columns = NULL
)
}
\arguments{
\item{data}{A dataframe containing the event locations in the columns x, y, and t. Defaults to NULL if not used.}

\item{params}{A named list of lists containing the values for the background rate, triggering ratio, spatial parameters in a named list, and temporal parameters in a named list. Defaults to NULL if not used.}

\item{time_window}{A numeric vector of length 2 specifying the time window.}

\item{spatial_region}{An sf object defining the spatial region and covariate regions.}

\item{spatial_family}{A spatial triggering kernel function to generate data from. Defaults to NULL if not used.}

\item{temporal_family}{A spatial triggering kernel function to generate data from. Defaults to NULL if not used.}

\item{covariate_columns}{A character vector containing the names of covariates columns to be used for the model.}
}
\value{
A hawkes object containing a tibble with the events.
}
\description{
Constructor for a hawkes object
}
