% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_likelihood_estimation.R
\name{hawkes_mle}
\alias{hawkes_mle}
\title{Function to Estimate MLEs}
\usage{
hawkes_mle(hawkes, inits, boundary = NULL, max_iters = 500)
}
\arguments{
\item{hawkes}{A \code{hawkes} object}

\item{inits}{A named list of lists containing the initial values for the background rate, triggering ratio, spatial parameters in a named list, and temporal parameters in a named list. Note that these values are of the same form as the true values in the Hawkes object but are often estimates passed to the function. If some parameters in the kernel function should be fixed in estimation, pass their names as another named list within params as fixed$spatial = c("mean").}

\item{boundary}{A boundary region to correct for the bpundary bias. Defaults to NULL if not used.}

\item{max_iters}{A numeric value for the maximum number of iteration in the EM-algorithm. Defaults to 500 if not used.}
}
\value{
A \code{hawkes_fit} object that is a list of lists contatiing the MLEs.
}
\description{
Function to Estimate MLEs
}
\examples{
set.seed(123)
region <- list(x = c(0,10), y = c(0,10), t = c(0,50))

params <- list(background_rate = list(intercept = -4),triggering_rate = 0.5,spatial = list(mean = 0, sd = 0.1),temporal = list(rate = 2), fixed = list(spatial = "mean", temporal = NULL))
hawkes <- rHawkes(params, region)
hawkes_mle(hawkes, inits = params, boundary = c(.5, 5))


data("example_background_covariates")
params <- list(background_rate = list(intercept = -4.5, X1 = 1, X2 = 1, X3 = 1),triggering_rate = 0.5,spatial = list(mean = 0, sd = .1),temporal = list(rate = 2), fixed = list(spatial = "mean"))
hawkes <- rHawkes(params, region, spatial_family = "Gaussian", temporal_family = "Exponential", cov_map = example_background_covariates)
est <- hawkes_mle(hawkes, inits = params, boundary = c(.5, 3))
}
