% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_trigger_lrt.R
\name{hawkes_trigger_lrt}
\alias{hawkes_trigger_lrt}
\title{Likelihood Ratio Test to Detect Triggering}
\usage{
hawkes_trigger_lrt(hawkes, parameters, alpha = 0.05)
}
\arguments{
\item{hawkes}{A \code{hawkes} object.}

\item{parameters}{Parameter values stored in a named list or a \code{hawkes_fit} object from \code{hawkes_mle()}.}

\item{alpha}{Type-1 error rate for the hypothesis test. Defaults to 0.05 if unused.}
}
\value{
A tibble
}
\description{
Likelihood Ratio Test to Detect Triggering
}
\examples{
spatial_region <- create_rectangular_sf(0, 10, 0, 10)

params <- list(
  background_rate = list(intercept = -4),
  triggering_rate = 0.01,
  spatial = list(mean = 0, sd = 0.5),
  temporal = list(rate = 1)
)
hawkes <- rHawkes(
  params = params,
  time_window = c(0, 100),
  spatial_region = spatial_region,
  background_process = ~ 1
)
est <- hawkes_mle(hawkes, ~ 1, inits = params)
hawkes_trigger_lrt(hawkes, est)


params <- list(
  background_rate = list(intercept = -4.5, X1 = 1, X2 = 1),
  triggering_rate = 0.1,
  spatial = list(mean = 0, sd = 0.25),
  temporal = list(rate = 2),
  fixed = list(spatial = "mean")
)
data("example_background_covariates")
hawkes <- rHawkes(
  params = params,
  time_window = c(0, 50),
  spatial_region = example_background_covariates,
  background_process = ~ X1 + X2,
  spatial_burnin = 1
)
est <- hawkes_mle(hawkes, ~ X1 + X2, inits = params, boundary = 1)
hawkes_trigger_lrt(hawkes, est)

}
