% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_likelihood_estimation.R
\name{hessian_est}
\alias{hessian_est}
\title{Estimate the covariance matrix of the MLE using the observed information}
\usage{
hessian_est(hawkes, est)
}
\arguments{
\item{hawkes}{An object containing the data and model specification for a Hawkes process,
passed to the log-likelihood function.}

\item{est}{A list of estimated parameter values (typically the output of an EM or MLE routine),
used as the starting point for computing the Hessian. Must match the structure expected by
\code{.flatten_free_params()} and \code{.vector_input_log_likelihood()}.}
}
\value{
A named covariance matrix corresponding to the estimated parameters.
}
\description{
Computes the observed information matrix (negative Hessian of the log-likelihood)
and returns its inverse as an estimate of the covariance matrix for the MLE parameters
in a spatio-temporal Hawkes process model.
}
\details{
This function requires the \pkg{numDeriv} package. It numerically approximates the
Hessian of the full log-likelihood using central differences and inverts the negative Hessian
to obtain the covariance estimate. The names of the matrix rows and columns are set to match
the flattened parameter vector.
}
\examples{
spatial_region <- create_rectangular_sf(0,10,0,10)

params <- list(background_rate = list(intercept = -4),triggering_rate = 0.5,spatial = list(mean = 0, sd = .1),temporal = list(rate = 2))
hawkes <- rHawkes(params, time_window = c(0,100), spatial_region = spatial_region)
est <- hawkes_mle(hawkes, inits = params)
hessian_est(hawkes, est$est)

}
