% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_fit.R
\name{new_hawkes_fit}
\alias{new_hawkes_fit}
\title{Construct a hawkes_fit object from estimated parameters}
\usage{
new_hawkes_fit(hawkes, est)
}
\arguments{
\item{hawkes}{The original \code{hawkes} object used to fit the model.}

\item{est}{Named nested list of estimated parameters, e.g., from \code{hawkes_mle()}.}
}
\value{
An object of class \code{hawkes_fit}.
}
\description{
Construct a hawkes_fit object from estimated parameters
}
