% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_bootstrap.R
\name{parametric_bootstrap}
\alias{parametric_bootstrap}
\title{Parametric bootstrap for confidence intervals of Hawkes MLEs}
\usage{
parametric_bootstrap(
  hawkes,
  est,
  B,
  alpha = 0.05,
  parallel = FALSE,
  max_iters = 500,
  boundary = NULL,
  temporal_burnin = NULL,
  spatial_burnin = NULL
)
}
\arguments{
\item{hawkes}{A \code{hawkes} object}

\item{est}{A \code{hawkes_mle} object to estimate the parameters for the \code{hawkes} object passed to the function.}

\item{B}{The number of bootstrap iterations. At least 1000 iterations are recommended, but computation time can be quite long.}

\item{alpha}{Specified alpha value. Defaults to 0.05 if not used.}

\item{parallel}{Logical to specify use of parallel computation via the \pkg{furrr} package. Setup multisession before calling function. See package for details.}

\item{max_iters}{A numeric value for the maximum number of iteration in the EM-algorithm. Defaults to 500 if not used.}

\item{boundary}{A boundary region to correct for the bpundary bias. Defaults to NULL if not used.}

\item{temporal_burnin}{A numeric value to set the temporal burn-in region for simulation.}

\item{spatial_burnin}{A numeric value to set the spatial burn-in region for simulation.}
}
\value{
A \code{hawkes_fit} object that is a list of lists contatiing the MLEs.
}
\description{
Parametric bootstrap for confidence intervals of Hawkes MLEs
}
\examples{
params <- list(background_rate = list(intercept = -4.5, X1 = 1, X2 = 1),triggering_rate = 0.5,spatial = list(mean = 0, sd = .25),temporal = list(rate = 2), fixed = list(spatial = "mean"))
data("example_background_covariates")
hawkes <- rHawkes(params, c(0,50), example_background_covariates, covariate_columns = c("X1", "X2"), spatial_burnin = 1)
est <- hawkes_mle(hawkes, inits = params, boundary = 1)

# B is set to 5 to demonstrate the use without requiring significant computation. We recommend at least B=1000
parametric_bootstrap(hawkes, est, B = 5, boundary = c(.5,3))


params <- list(background_rate = list(intercept = -4.5, X1 = 1, X2 = 1),triggering_rate = 0.5,spatial = list(mean = 0, sd = .25),temporal = list(rate = 2), fixed = list(spatial = "mean"))
data("example_background_covariates")
hawkes <- rHawkes(params, c(0,50), example_background_covariates, covariate_columns = c("X1", "X2"), spatial_burnin = 1)
est <- hawkes_mle(hawkes, inits = params, boundary = 1)

future::plan(future::multisession, workers = future::availableCores())

parametric_bootstrap(hawkes, est, B = 2, parallel = TRUE, boundary = c(.5,3))

future::plan(future::sequential)
}
