% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_likelihood_estimation.R
\name{parent_est}
\alias{parent_est}
\title{Parent Matrix Estimation}
\usage{
parent_est(hawkes, parameters)
}
\arguments{
\item{hawkes}{A \code{hawkes} object.}

\item{parameters}{Named list containing background, triggering, spatial, and temporal
parameters. Values usually represent current estimates.}
}
\value{
A numeric matrix
}
\description{
Parent Matrix Estimation
}
