% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_likelihood_estimation.R
\name{parent_est}
\alias{parent_est}
\title{Parent Matrix Estimation}
\usage{
parent_est(hawkes, parameters)
}
\arguments{
\item{hawkes}{A \code{hawkes} object.}

\item{parameters}{Named list containing background, triggering, spatial, and temporal
parameters. Values usually represent current estimates.}
}
\value{
A numeric matrix
}
\description{
Parent Matrix Estimation
}
\examples{
spatial_region <- create_rectangular_sf(0, 10, 0, 10)

params <- list(
  background_rate = list(intercept = -4),
  triggering_rate = 0.75,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2)
)
hawkes <- rHawkes(
  params = params,
  time_window = c(0, 50),
  spatial_region = spatial_region,
  background_process = ~ 1
)

(parent_est_mat <- parent_est(hawkes, params))

params <- list(
  background_rate = list(intercept = -4.5, X1 = 1, X2 = 1),
  triggering_rate = 0.5,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2),
  fixed = list(spatial = "mean")
)
data("example_background_covariates")
hawkes <- rHawkes(
  params = params,
  time_window = c(0, 50),
  spatial_region = example_background_covariates,
  background_process = ~ X1 + X2,
  spatial_burnin = 0
)

(parent_est_mat <- parent_est(hawkes, params))

params <- list(
  background_rate = list(intercept = -4,
                         event_type = c(a = 1, b = .25, c = .5)),
  triggering_rate = matrix(c(.4, .15, .05,
                             .2, .05, .02,
                             .2, .05, .20),
                           nrow = 3,
                           dimnames = list(c("a", "b", "c"), c("a", "b", "c"))),
  spatial = list(mean = 0, sd = 0.1),
  temporal = list(rate = 2)
)
hawkes <- rHawkes(
  params = params,
  time_window = c(0, 50),
  spatial_region = spatial_region,
  background_process = ~ 1 + mark(event_type),
  spatial_burnin = 1
)

(parent_est_mat <- parent_est(hawkes, params))
}
