% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_likelihood_estimation.R
\name{parent_est}
\alias{parent_est}
\title{Parent Matrix Estimation}
\usage{
parent_est(hawkes, parameters)
}
\arguments{
\item{hawkes}{A \code{hawkes} object}

\item{parameters}{A named list of lists containing the values for the background rate, triggering ratio, spatial parameters in a named list, and temporal parameters in a named list. Note that these values are of the same form as the true values in the Hawkes object but are often estimates passed to the function.}
}
\value{
A numeric matrix
}
\description{
Parent Matrix Estimation
}
\examples{
spatial_region <- create_rectangular_sf(0,10,0,10)

params <- list(background_rate = list(intercept = -4),triggering_rate = 0.75,spatial = list(mean = 0, sd = .75),temporal = list(rate = 2))
hawkes <- rHawkes(params, time_window = c(0,50), spatial_region = spatial_region)

(parent_est_mat <- parent_est(hawkes, params))

params <- list(background_rate = list(intercept = -4.5, X1 = 1, X2 = 1),triggering_rate = 0.5,spatial = list(mean = 0, sd = .75),temporal = list(rate = 2), fixed = list(spatial = "mean"))
data("example_background_covariates")
hawkes <- rHawkes(params, c(0,50), example_background_covariates, covariate_columns = c("X1", "X2"), spatial_burnin = 0)
}
