% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_likelihood_estimation.R
\name{parent_est}
\alias{parent_est}
\title{Parent Matrix Estimation}
\usage{
parent_est(hawkes, parameters = NULL)
}
\arguments{
\item{hawkes}{A \code{hawkes} object}

\item{parameters}{A named list of lists containing the values for the background rate, triggering ratio, spatial parameters in a named list, and temporal parameters in a named list. Note that these values are of the same form as the true values in the Hawkes object but are often estimates passed to the function.}
}
\value{
A numeric matrix
}
\description{
Parent Matrix Estimation
}
\examples{
region <- list(x = c(0,10), y = c(0,10), t = c(0,50))

params <- list(background_rate = list(intercept = -4),triggering_rate = 0.5,spatial = list(mean = 0, sd = 0.5),temporal = list(rate = 2), fixed = list(spatial = "mean", temporal = NULL))
hawkes <- rHawkes(params, region)
(parent_est_mat <- parent_est(hawkes, params))

}
