% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hawkes.R
\name{plot_hawkes}
\alias{plot_hawkes}
\title{Plot Hawkes Process Points}
\usage{
plot_hawkes(hawkes, color = "time", ...)
}
\arguments{
\item{hawkes}{A \code{hawkes} object}

\item{color}{Either \code{"time"} to color by event time or \code{"background"} to show
simulated background events.}

\item{...}{Additional arguments forwarded to \code{ggplot2::geom_sf()}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Creates a scatter plot of Hawkes events colored by time or background status.
}
\examples{
spatial_region <- create_rectangular_sf(0, 10, 0, 10)

params <- list(
  background_rate = list(intercept = -4),
  triggering_rate = 0.75,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2)
)
hawkes <- rHawkes(
  params,
  time_window = c(0, 50),
  spatial_region = spatial_region,
  spatial_burnin = 1
)

plot_hawkes(hawkes, color = "time")


params <- list(
  background_rate = list(intercept = -4.5, X1 = 1, X2 = 1),
  triggering_rate = 0.5,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2),
  fixed = list(spatial = "mean")
)
data("example_background_covariates")
hawkes <- rHawkes(
  params = params,
  time_window = c(0, 50),
  spatial_region = example_background_covariates,
  background_process = ~ X1 + X2,
  spatial_burnin = 1
)

plot_hawkes(hawkes, color = "time")
}
