% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hawkes.R
\name{plot_hawkes}
\alias{plot_hawkes}
\title{Plot Hawkes Process Points}
\usage{
plot_hawkes(hawkes, color = "time", ...)
}
\arguments{
\item{hawkes}{A \code{hawkes} object}

\item{color}{A character string. Either \code{"time"} to color by time, or \code{"background"} to color by whether an event is a background event (only works for simulated processes).}

\item{...}{Additional arguments passed to \code{ggplot2::geom_point()}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Creates a scatterplot of points from a spatio-temporal Hawkes process, colored either by time or by whether they are background events.
}
\examples{
# example code
set.seed(123)
region <- list(x = c(0,10), y = c(0,10), t = c(0,50))

params <- list(background_rate = list(intercept = -4),triggering_rate = 0.5,spatial = list(mean = 0, sd = 0.1),temporal = list(rate = 2), fixed = list(spatial = "mean", temporal = NULL))
hawkes <- rHawkes(params, region)
est <- hawkes_mle(hawkes, inits = params, boundary = c(.5, 5))
plot_hawkes(hawkes)

data("example_background_covariates")
params <- list(background_rate = list(intercept = -4.5, X1 = 1, X2 = 1, X3 = 1),triggering_rate = 0.5,spatial = list(mean = 0, sd = .1),temporal = list(rate = 2), fixed = list(spatial = "mean"))
hawkes <- rHawkes(params, region, spatial_family = "Gaussian", temporal_family = "Exponential", cov_map = example_background_covariates)
est <- hawkes_mle(hawkes, inits = params, boundary = c(.5, 3))
plot_hawkes(hawkes, color = "time")
}
