% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hawkes.R
\name{plot_intensity}
\alias{plot_intensity}
\title{Plot conditional intensity of a Hawkes process}
\usage{
plot_intensity(hawkes, est, stepsize, time = NULL, coordinates = NULL)
}
\arguments{
\item{hawkes}{A \code{hawkes} object}

\item{est}{A list of fitted parameter values for the Hawkes process.}

\item{stepsize}{A numeric value for the spatial resolution of the evaluation grid.}

\item{time}{A numeric value giving the time at which to evaluate the conditional intensity.}

\item{coordinates}{Numeric vector of length two giving the evaluation location.}
}
\value{
A list of \code{ggplot2} objects showing spatial and temporal intensity views.
}
\description{
Visualizes the conditional intensity over space or time given a fitted Hawkes model.
}
\examples{
# example code
set.seed(123)
spatial_region <- create_rectangular_sf(0, 10, 0, 10)

params <- list(
  background_rate = list(intercept = -4),
  triggering_rate = 0.75,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2)
)
hawkes <- rHawkes(params = params, time_window = c(0, 50), spatial_region = spatial_region)
est <- hawkes_mle(hawkes, inits = params, boundary = 1)

plot_intensity(hawkes, est, stepsize = 0.25, time = 40)
plot_intensity(hawkes, est, stepsize = 0.25, coordinates = c(5, 5))
plot_intensity(hawkes, est, stepsize = 0.25, coordinates = c(5, 5), time = 40)

params <- list(
  background_rate = list(intercept = -4.5, X1 = 1, X2 = 1),
  triggering_rate = 0.5,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2),
  fixed = list(spatial = "mean")
)
data("example_background_covariates")
hawkes <- rHawkes(
  params = params,
  time_window = c(0, 50),
  spatial_region = example_background_covariates,
  background_process = ~ X1 + X2,
  spatial_burnin = 1
)
est <- hawkes_mle(hawkes, inits = params, boundary = c(.5, 3))
plot_hawkes(hawkes)
plot_intensity(hawkes, est, stepsize = 0.1, time = 40, coordinates = c(4.5, 5))
}
