% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hawkes.R
\name{plot_intensity}
\alias{plot_intensity}
\title{Plot Conditional Intensity of Hawkes Process}
\usage{
plot_intensity(hawkes, est, stepsize, time = NULL, coordinates = NULL)
}
\arguments{
\item{hawkes}{A \code{hawkes} object}

\item{est}{A list of fitted parameter values for the Hawkes process.}

\item{stepsize}{A numeric value for the spatial resolution of the evaluation grid.}

\item{time}{A numeric value giving the time at which to evaluate the conditional intensity.}

\item{coordinates}{A numeric vector of length 2 specifying the coordinates where to evalute the conditional intensity over time.}
}
\value{
A list of \code{ggplot2} objects showing the spatial intensity and temporal intensity at a given time or location.
}
\description{
Visualizes the conditional intensity function at a fixed time over a spatial grid, based on a fitted Hawkes model.
}
\examples{
# example code
set.seed(123)
region <- list(x = c(0,10), y = c(0,10), t = c(0,50))

params <- list(background_rate = list(intercept = -4),triggering_rate = 0.5,spatial = list(mean = 0, sd = 0.25),temporal = list(rate = 1), fixed = list(spatial = "mean", temporal = NULL))
hawkes <- rHawkes(params, region)
est <- hawkes_mle(hawkes, inits = params, boundary = c(.5, 5))
plot_intensity(hawkes, est, time = 40)
plot_intensity(hawkes, est, coordinates = c(5,5))
plot_intensity(hawkes, est, coordinates = c(5,5), time = 40)

data("example_background_covariates")
params <- list(background_rate = list(intercept = -3.5, X1 = 1, X2 = .75, X3 = .5),triggering_rate = 0.5,spatial = list(mean = 0, sd = .25),temporal = list(rate = 1), fixed = list(spatial = "mean"))
hawkes <- rHawkes(params, region, spatial_family = "Gaussian", temporal_family = "Exponential", cov_map = example_background_covariates)
est <- hawkes_mle(hawkes, inits = params, boundary = c(.5, 3))
plot_hawkes(hawkes)
plot_intensity(hawkes, est, time = 40, coordinates = c(4.5, 5))
}
