% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rHawkes.R
\name{rHawkes}
\alias{rHawkes}
\title{Generate a Hawkes Process}
\usage{
rHawkes(
  params,
  region,
  t_burnin = 50,
  s_burnin = 0,
  cov_map = NULL,
  spatial_kernel = rnorm,
  temporal_kernel = rexp
)
}
\arguments{
\item{params}{A named list of lists containing the values for the background rate, triggering ratio, spatial parameters in a named list, and temporal parameters in a named list.}

\item{region}{A list containing the spatial and temporal windows of the form list(x = c(xmin, xmax), y = c(ymin, ymax), t = c(tmin, tmax)).}

\item{t_burnin}{Temporal burn-in for simulation. Defaults to 50 if not used.}

\item{s_burnin}{Spatial burn-in for simulation. Note that this may not work as expected with a background covariate map. Defaults to 0 if not used.}

\item{cov_map}{An sf object containing spatial polygons with associated covariate values named X1, X2, .... Defaults to NULL if not used.}

\item{spatial_kernel}{A spatial triggering kernel function to generate data from. Defaults to rnorm if not used.}

\item{temporal_kernel}{A spatial triggering kernel function to generate data from. Defaults to rexp if not used.}
}
\value{
A hawkes object with a generated Hawkes process.
}
\description{
Generate a Hawkes Process
}
\examples{
region <- list(x = c(0,10), y = c(0,10), t = c(0,50))

params <- list(background_rate = -4,triggering_rate = 0.5,spatial = list(mean = 0, sd = 0.5),temporal = list(rate = 2))
rHawkes(params, region)

params <- list(background_rate = -4,triggering_rate = 0.5,spatial = list(min = -1, max = 1),temporal = list(shape = 3, rate = 1))
rHawkes(params, region, spatial_kernel = runif, temporal_kernel = rgamma)

params <- list(background_rate = -4,triggering_rate = 0.5,spatial = list(rate = 5),temporal = list(rate = 2))
rHawkes(params, region, spatial_kernel = rexp_spatial, temporal_kernel = rexp)


#params <- list(background_rate = c(-4.5, 1),triggering_rate = 0.5,spatial = list(rate = 5),temporal = list(rate = 2))
#data("example_background_covariates")
#example_background_covariates <- example_background_covariates[,c("geoid", "name", "area", "X1")]
#rHawkes(params, region, spatial_kernel = rexp_spatial, temporal_kernel = rexp, cov_map = example_background_covariates)
}
