% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rHawkes.R
\name{rHawkes}
\alias{rHawkes}
\title{Generate a Hawkes Process}
\usage{
rHawkes(
  params,
  time_window,
  spatial_region,
  covariate_columns = NULL,
  temporal_burnin = (time_window[2] - time_window[1])/(10),
  spatial_burnin = sum(sf::st_area(spatial_region))^0.25,
  spatial_family = "Gaussian",
  temporal_family = "Exponential"
)
}
\arguments{
\item{params}{A named list of lists containing the values for the background rate, triggering ratio, spatial parameters in a named list, and temporal parameters in a named list.}

\item{time_window}{A numeric vector of length 2 specifying the simulated time window.}

\item{spatial_region}{A list containing the spatial and temporal windows of the form list(x = c(xmin, xmax), y = c(ymin, ymax), t = c(tmin, tmax)).}

\item{covariate_columns}{A character vector of the names of the columns in spatial_region to be used as background covariates.}

\item{spatial_family}{A spatial triggering kernel function to generate data from. Alternatively, a list can be provided to designate a custom kernel. The list must contain the objects named spatial_pdf, spatial_cdf, and spatial_sampler. They should follow the format of the dnorm, pnorm, and rnorm functions and the parameters must match the names. Defaults to NULL if not used.}

\item{temporal_family}{A spatial triggering kernel function to generate data from. Alternatively, a list can be provided to designate a custom kernel. The list must contain the objects named temporal_pdf, temporal_cdf, and temporal_sampler. They should follow the format of the dnorm, pnorm, and rnorm functions and the parameters must match the names. Defaults to NULL if not used.}

\item{t_burnin}{Temporal burn-in for simulation. Defaults to 10 if not used.}

\item{s_burnin}{Spatial burn-in for simulation. Note that this may not work as expected with a background covariate map. Defaults to 0 if not used.}
}
\value{
A hawkes object with a generated Hawkes process.
}
\description{
Generate a Hawkes Process
}
\examples{
spatial_region <- create_rectangular_sf(0,10,0,10)

params <- list(background_rate = list(intercept = -4),triggering_rate = 0.75,spatial = list(mean = 0, sd = .75),temporal = list(rate = 2))
rHawkes(params, time_window = c(0,50), spatial_region = spatial_region)

params <- list(background_rate = list(intercept = -4.5, X1 = 1, X2 = 1),triggering_rate = 0.5,spatial = list(mean = 0, sd = .75),temporal = list(rate = 2), fixed = list(spatial = "mean"))
data("example_background_covariates")
rHawkes(params, c(0,50), example_background_covariates, covariate_columns = c("X1", "X2"), spatial_burnin = 0)
}
