% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rHawkes.R
\name{rHawkes}
\alias{rHawkes}
\title{Generate a Hawkes Process}
\usage{
rHawkes(
  params,
  time_window,
  spatial_region,
  covariate_columns = NULL,
  temporal_burnin = (time_window[2] - time_window[1])/(10),
  spatial_burnin = sum(as.numeric(sf::st_area(spatial_region)))^0.25,
  temporal_family = "Exponential",
  spatial_family = "Gaussian"
)
}
\arguments{
\item{params}{A named list of lists containing the values for the background rate, triggering rate, spatial parameters in a named list, and temporal parameters in a named list. See example for exact specification.}

\item{time_window}{A numeric vector of length 2 specifying the simulated time window.}

\item{spatial_region}{An sf object defining the spatial region for simulation.}

\item{covariate_columns}{A character vector of the names of the columns in spatial_region to be used as background covariates.}

\item{temporal_burnin}{Temporal burn-in for simulation. Defaults to 1/10 of the length of the time window if not used.}

\item{spatial_burnin}{Spatial burn-in for simulation. Defaults to the area of spatial_region^.25.}

\item{temporal_family}{A temporal triggering kernel function to generate data from. Defaults to "Exponential" if not used. Other options include "Power Law", "Uniform", and "Gamma".}

\item{spatial_family}{A spatial triggering kernel function to generate data from. Defaults to "Gaussian" if not used. Other options include "Uniform" and "Exponential"}
}
\value{
A hawkes object with a generated Hawkes process.
}
\description{
Generate a Hawkes Process
}
\examples{
spatial_region <- create_rectangular_sf(0,10,0,10)

params <- list(background_rate = list(intercept = -4),triggering_rate = 0.75,spatial = list(mean = 0, sd = .75),temporal = list(rate = 2))
(hawkes <- rHawkes(params, time_window = c(0,50), spatial_region = spatial_region, spatial_burnin = 1))

params <- list(background_rate = list(intercept = -4.5, X1 = 1, X2 = 1),triggering_rate = 0.5,spatial = list(mean = 0, sd = .75),temporal = list(rate = 2), fixed = list(spatial = "mean"))
data("example_background_covariates")
rHawkes(params, c(0,50), example_background_covariates, covariate_columns = c("X1", "X2"), spatial_burnin = 1)
}
