% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_bootstrap.R
\name{sample_blocks}
\alias{sample_blocks}
\title{Resample data with moving blocls}
\usage{
sample_blocks(hawkes, num_blocks)
}
\arguments{
\item{hawkes}{A \code{hawkes} object}

\item{block_length_t}{A numeric value to set the length of blocks to wrap the process.}
}
\value{
A \code{hawkes} object.
}
\description{
Resample data with moving blocls
}
\examples{
region <- list(x = c(0,10), y = c(0,10), t = c(0,50))

params <- list(background_rate = list(intercept = -4),triggering_rate = 0.5,spatial = list(mean = 0, sd = 0.1),temporal = list(rate = 2), fixed = list(spatial = "mean", temporal = NULL))
hawkes <- rHawkes(params, region)
sample_blocks(hawkes, 25)

}
