% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_bootstrap.R
\name{sample_blocks}
\alias{sample_blocks}
\title{Resample data with moving blocls}
\usage{
sample_blocks(hawkes, num_blocks)
}
\arguments{
\item{hawkes}{A \code{hawkes} object}

\item{block_length_t}{A numeric value to set the length of blocks to wrap the process.}
}
\value{
A \code{hawkes} object.
}
\description{
Resample data with moving blocls
}
\examples{
params <- list(background_rate = list(intercept = -4.5, X1 = 1, X2 = 1),triggering_rate = 0.5,spatial = list(mean = 0, sd = .25),temporal = list(rate = 2), fixed = list(spatial = "mean"))
data("example_background_covariates")
hawkes <- rHawkes(params, c(0,50), example_background_covariates, covariate_columns = c("X1", "X2"), spatial_burnin = 1)

sample_blocks(hawkes, 25)

}
