% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_bootstrap.R
\name{sample_clusters}
\alias{sample_clusters}
\title{Resample Data by EM-ALgorithm Clustering}
\usage{
sample_clusters(hawkes, parent_mat, boundary = NULL)
}
\arguments{
\item{hawkes}{A \code{hawkes} object}

\item{parent_mat}{An estimate of the parent matrix produced by \code{parent_est}}
}
\value{
A \code{hawkes} object.
}
\description{
Resample Data by EM-ALgorithm Clustering
}
\examples{
region <- list(x = c(0,10), y = c(0,10), t = c(0,50))

params <- list(background_rate = list(intercept = -4),triggering_rate = 0.5,spatial = list(mean = 0, sd = 0.1),temporal = list(rate = 2), fixed = list(spatial = "mean", temporal = NULL))
hawkes <- rHawkes(params, region)
est <- hawkes_mle(hawkes, inits = params, boundary = c(.5, 3))
parent_mat <- parent_est(hawkes, est)
sample_clusters(hawkes, parent_mat, boundary = c(.5,3))

}
