% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_bootstrap.R
\name{sample_clusters}
\alias{sample_clusters}
\title{Resample Data by EM-ALgorithm Clustering}
\usage{
sample_clusters(hawkes, parent_mat, boundary = NULL)
}
\arguments{
\item{hawkes}{A \code{hawkes} object.}

\item{parent_mat}{An estimate of the parent matrix produced by \code{parent_est}.}

\item{boundary}{size of boundary to use for border correction. Defaults to NULL if unused.}
}
\value{
A \code{hawkes} object.
}
\description{
Resample Data by EM-ALgorithm Clustering
}
\examples{
params <- list(background_rate = list(intercept = -4.5, X1 = 1, X2 = 1),triggering_rate = 0.5,spatial = list(mean = 0, sd = .25),temporal = list(rate = 2), fixed = list(spatial = "mean"))
data("example_background_covariates")
hawkes <- rHawkes(params, c(0,50), example_background_covariates, covariate_columns = c("X1", "X2"), spatial_burnin = 1)
est <- hawkes_mle(hawkes, inits = params, boundary = 1)
parent_mat <- parent_est(hawkes, est)
sample_clusters(hawkes, parent_mat, boundary = c(.5,3))

}
