% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{sample_subregion}
\alias{sample_subregion}
\title{Resample data with moving blocls}
\usage{
sample_subregion(hawkes, length)
}
\arguments{
\item{hawkes}{A \code{hawkes} object}

\item{length}{A numeric value to set the length of the subsample.}
}
\value{
A \code{hawkes} object.
}
\description{
Resample data with moving blocls
}
\examples{
region <- list(x = c(0,10), y = c(0,10), t = c(0,50))

params <- list(background_rate = list(intercept = -4),triggering_rate = 0.5,spatial = list(mean = 0, sd = 0.1),temporal = list(rate = 2), fixed = list(spatial = "mean", temporal = NULL))
hawkes <- rHawkes(params, region)
sample_subregion(hawkes, 25)

}
