% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rHawkes.R
\name{sim_background_events}
\alias{sim_background_events}
\title{Simulate background events}
\usage{
sim_background_events(
  background_rate,
  time_window,
  spatial_region,
  covariate_columns = NULL
)
}
\arguments{
\item{background_rate}{Vector of coefficients for the background covariates.}

\item{time_window}{Numeric vector of length two giving the simulated time window.}

\item{spatial_region}{\code{sf} object defining the simulation region.}

\item{covariate_columns}{Optional character vector naming background covariates in
\code{spatial_region}.}
}
\value{
A dataframe with generated background events.
}
\description{
Simulate background events
}
\examples{
spatial_region <- create_rectangular_sf(0, 10, 0, 10)
time_window <- c(0, 50)
background_rate <- -4

sim_background_events(background_rate, time_window, spatial_region)
}
