% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rHawkes.R
\name{sim_background_events}
\alias{sim_background_events}
\title{Simulate background events}
\usage{
sim_background_events(background_rate, region, cov_map = NULL)
}
\arguments{
\item{background_rate}{A vector of coefficients for the background covariates.}

\item{region}{A list containing the spatial and temporal windows of the form list(x = c(xmin, xmax), y = c(ymin, ymax), t = c(tmin, tmax)).}

\item{cov_map}{An sf object containing background covariates.}
}
\value{
A hawkes object with only generated background events.
}
\description{
Simulate background events
}
\examples{
region <- list(x = c(0,10), y = c(0,10), t = c(0,50))
background_rate <- -4

sim_background_events(background_rate, region)
}
