% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rHawkes.R
\name{sim_background_events}
\alias{sim_background_events}
\title{Simulate background events}
\usage{
sim_background_events(
  background_rate,
  time_window,
  spatial_region,
  cov_map = NULL
)
}
\arguments{
\item{background_rate}{A vector of coefficients for the background covariates.}

\item{cov_map}{An sf object containing background covariates.}

\item{region}{A list containing the spatial and temporal windows of the form list(x = c(xmin, xmax), y = c(ymin, ymax), t = c(tmin, tmax)).}
}
\value{
A dataframe with generated background events.
}
\description{
Simulate background events
}
\examples{
spatial_region <- create_rectangular_sf(0,10,0,10)
time_window <- c(0,50)
background_rate <- -4

sim_background_events(background_rate, time_window, spatial_region)
}
