% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{spatial_conditional_intensity}
\alias{spatial_conditional_intensity}
\title{Compute the conditional intensity at a specified time}
\usage{
spatial_conditional_intensity(hawkes, parameters, time, stepsize)
}
\arguments{
\item{hawkes}{A \code{hawkes} object.}

\item{parameters}{Named list containing background, triggering, spatial, and temporal
parameters.}

\item{time}{Time point where the conditional intensity is evaluated.}

\item{stepsize}{Grid cell size used to evaluate the spatial intensity.}
}
\value{
A numeric vector.
}
\description{
Compute the conditional intensity at a specified time
}
\examples{
spatial_region <- create_rectangular_sf(0, 10, 0, 10)

params <- list(
  background_rate = list(intercept = -4),
  triggering_rate = 0.75,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2)
)
hawkes <- rHawkes(params = params, time_window = c(0, 50), spatial_region = spatial_region)
spatial_conditional_intensity(hawkes, params, 25, 0.5)

params <- list(
  background_rate = list(intercept = -4.5, X1 = 1, X2 = 1),
  triggering_rate = 0.5,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2),
  fixed = list(spatial = "mean")
)
data("example_background_covariates")
hawkes <- rHawkes(
  params = params,
  time_window = c(0, 50),
  spatial_region = example_background_covariates,
  background_process = ~ X1 + X2,
  spatial_burnin = 1
)

spatial_conditional_intensity(hawkes, params, 25, 0.5)
}
