% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample}
\alias{subsample}
\title{Block Bootstrap for COnfidence Intervals of Hawkes MLEs}
\usage{
subsample(
  hawkes,
  est,
  B,
  length,
  alpha,
  parallel = FALSE,
  max_iters = 500,
  boundary = NULL
)
}
\arguments{
\item{hawkes}{A \code{hawkes} object}

\item{est}{A \code{hawkes_est} object containing the parameter estimates of \code{hawkes}}

\item{B}{number of bootstrap iterations}

\item{length}{A numeric value to set the length of the subsample}

\item{alpha}{type-1 error rate for constructing confidence intervals. Defaults to .05 if unused}

\item{parallel}{a logical specifying if parallel computation should be used. Parallel computation is implemented with the \code{furrr} package.}

\item{max_iters}{maximum number of iterations to use in maximum likelihood estimation. Defaults to 500 if unused.}

\item{boundary}{size of boundary to use for border correction. Defaults to NULL if unused}
}
\value{
A \code{hawkes} object.
}
\description{
Block Bootstrap for COnfidence Intervals of Hawkes MLEs
}
\examples{
spatial_region <- create_rectangular_sf(0,10,0,10)

params <- list(background_rate = list(intercept = -4),triggering_rate = 0.75,spatial = list(mean = 0, sd = .75),temporal = list(rate = 2))
hawkes <- rHawkes(params, time_window = c(0,50), spatial_region = spatial_region)
est <- hawkes_mle(hawkes, inits = params)
subsample(hawkes, est, 5, 25, alpha = 0.05)
}
