% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample}
\alias{subsample}
\title{Block bootstrap for Hawkes MLEs via subsampling}
\usage{
subsample(
  hawkes,
  est,
  B,
  length,
  alpha,
  parallel = FALSE,
  max_iters = 500,
  boundary = NULL
)
}
\arguments{
\item{hawkes}{A \code{hawkes} object.}

\item{est}{A \code{hawkes_est} object containing parameter estimates.}

\item{B}{Number of bootstrap iterations.}

\item{length}{Length of each temporal subsample.}

\item{alpha}{Type I error rate for confidence intervals. Defaults to 0.05.}

\item{parallel}{Logical flag for \code{furrr} based parallel computation.}

\item{max_iters}{Maximum EM iterations per fit. Defaults to 500.}

\item{boundary}{Optional boundary width used for edge correction.}
}
\value{
A tibble of bootstrap estimates.
}
\description{
Block bootstrap for Hawkes MLEs via subsampling
}
\examples{
spatial_region <- create_rectangular_sf(0, 10, 0, 10)

params <- list(
  background_rate = list(intercept = -4),
  triggering_rate = 0.75,
  spatial = list(mean = 0, sd = 0.75),
  temporal = list(rate = 2)
)
hawkes <- rHawkes(params, time_window = c(0, 50), spatial_region = spatial_region)
est <- hawkes_mle(hawkes, inits = params)
subsample(hawkes, est, 5, 25, alpha = 0.05)
}
