% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{temporal_conditional_intensity}
\alias{temporal_conditional_intensity}
\title{Compute the Conditional Intensity at a Specified Location (x,y)}
\usage{
temporal_conditional_intensity(hawkes, parameters, coordinates, step = 0.1)
}
\arguments{
\item{hawkes}{A \code{hawkes} object.}

\item{parameters}{A named list of lists containing the values for the background rate, triggering ratio, spatial parameters in a named list, and temporal parameters in a named list. Note that these values are of the same form as the true values in the Hawkes object but are often estimates passed to the function.}

\item{coordinates}{A numeric vector of length 2 to specify the location to evaluate the conditional intensity.}

\item{step}{A numeric value specifying the size of the grid to evaluate the conditional intensity over.}
}
\value{
A numeric vector
}
\description{
Compute the Conditional Intensity at a Specified Location (x,y)
}
\examples{
region <- list(x = c(0,10), y = c(0,10), t = c(0,50))

params <- list(background_rate = -4,triggering_rate = 0.5,spatial = list(mean = 0, sd = 0.5),temporal = list(rate = 2))
hawkes <- rHawkes(params, region)
temporal_conditional_intensity(hawkes, params, c(5,5))
}
