% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{time_scaled_residuals}
\alias{time_scaled_residuals}
\title{Compute Time Scaled Residuals}
\usage{
time_scaled_residuals(hawkes, est)
}
\arguments{
\item{hawkes}{a \code{hawkes} object}

\item{est}{a \code{hawkes_mle} object produced from \code{hawkes_mle()}.}
}
\value{
A numeric vector of transformed waiting times.
}
\description{
Compute Time Scaled Residuals
}
\examples{
region <- list(x = c(0,10), y = c(0,10), t = c(0,50))

params <- list(background_rate = list(intercept = -4),triggering_rate = 0.5,spatial = list(mean = 0, sd = 0.1),temporal = list(rate = 2), fixed = list(spatial = "mean", temporal = NULL))
hawkes <- rHawkes(params, region)
est <- hawkes_mle(hawkes, inits = params, boundary = c(.5, 5))
residuals <- time_scaled_residuals(hawkes, est)

plot(density(rexp(length(residuals)), from = 0), lwd = 2)
lines(density(residuals, from = 0), col = "red")

params <- list(background_rate = list(intercept = -4),triggering_rate = 0.5,spatial = list(min = -1, max = 1),temporal = list(shape = 3, rate = 1))
hawkes <- rHawkes(params, region, spatial_family = "Uniform", temporal_family = "Gamma")
est <- hawkes_mle(hawkes, inits = params, boundary = c(.5, 5))
residuals <- time_scaled_residuals(hawkes, est)

plot(density(rexp(length(residuals)), from = 0), lwd = 2)
lines(density(residuals, from = 0), col = "red")
}
